Code complete with parameterized test

+++ src/Fibonacci.java
@@ -4,22 +4,22 @@
 
 public class Fibonacci {
 
-    private static int[] precalculated = null;
+    private static long[] precalculated = null;
 
-    public static int fibImproved(int n) {
+    public static long fibImproved(int n) {
         if (precalculated == null) {
             initPrecalculatedArray(n);
         }
         if (precalculated[n-1] != -1) {
             return precalculated[n-1];
         } else {
-            int result = fibImproved(n-1) + fibImproved(n-2);
+            long result = fibImproved(n-1) + fibImproved(n-2);
             precalculated[n-1] = result;
             return result;
         }
     }
     private static void initPrecalculatedArray(int size) {
-        precalculated = new int[size];
+        precalculated = new long[size];
         for (int i = 0; i < precalculated.length; i++) {
             precalculated[i] = -1; // to indicate "not calculated yet"
         }
@@ -27,11 +27,11 @@ public class Fibonacci {
         precalculated[1] = 1; // F(2)
     }
 
-    public static int fib(int n) {
+    public static long fib(int n) {
         if ((n == 1) || (n == 2)) {
             return 1;
         } else {
-            int result = fib(n-1) + fib(n-2);
+            long result = fib(n-1) + fib(n-2);
             return result;
         }
     }
